Prevent anaconda-core requiring gjs-console
Between 28.18-1 and 28.19-1, anaconda-core suddenly grew a
dependency on /usr/bin/gjs-console. This is ultimately due to
this change:
https://pagure.io/packaging-committee/issue/738
That thingy mungs the shebang in fedora-welcome.js from
/usr/bin/env gjs-console to /usr/bin/gjs-console, which in turn
causes RPM's automatic dependency code to make anaconda-core
depend on /usr/bin/gjs-console .

In a sense this is 'correct', but it's clearly problematic,
because it means we wind up depending on gjs (and gtk3!) in
all sorts of cases where we don't want to. This breaks stuff,
for instance arm minimal disk image composes. Ultimately perhaps
the GNOME welcome screen should be split out from anaconda-core,
but for now this will fix things up.

I have already made this change in dist-git and sent out the
28.20-2 build to Rawhide.